Update OpenSees3.3.0.iss
Drastic change to how OpenSees is installed. Found that OpenSees could be placed in a separate folder, and not the Tcl binary folder. Additionally, this allows for an OpenSees-specific lib folder, which is really cool.
Added the OpenSees batch file for convenience, since a lot of people use that method.
This methodology is great for having the full capabilities of Tcl while also having separate folders for certain OpenSees versions, while also giving the option to add the OpenSees binary to path.
Best of all worlds really.